refactor(JSON-RPC): split JsonRpcError for each error
