+++ crates/papyrus_gateway/src/api.rs
@@ -1,4 +1,3 @@
-use std::collections::HashSet;
 use std::sync::Arc;
 
 use jsonrpsee::{Methods, RpcModule};
@@ -6,8 +5,7 @@ use papyrus_common::BlockHashAndNumber;
 use papyrus_storage::StorageReader;
 use serde::{Deserialize, Serialize};
 use starknet_api::block::{BlockHash, BlockNumber};
-use starknet_api::core::{ChainId, ContractAddress};
-use starknet_api::transaction::EventKey;
+use starknet_api::core::ChainId;
 use starknet_client::writer::StarknetWriter;
 use tokio::sync::RwLock;
