+++ crates/papyrus_gateway/src/v0_4_0/api/api_impl.rs
@@ -28,15 +28,16 @@ use super::super::transaction::{
     TransactionWithHash, Transactions,
 };
 use super::{
-    BlockHashAndNumber, BlockId, EventFilter, EventsChunk, GatewayContractClass, JsonRpcV0_4Server,
+    BlockHashAndNumber, BlockId, ContinuationToken, EventFilter, EventsChunk, GatewayContractClass,
+    JsonRpcV0_4Server,
 };
-use crate::api::{BlockHashOrNumber, ContinuationToken, JsonRpcError, JsonRpcServerImpl};
-use crate::block::get_block_header_by_number;
+use crate::api::{BlockHashOrNumber, JsonRpcServerImpl};
 use crate::syncing_state::SyncingState;
-use crate::transaction::{get_block_tx_hashes_by_number, get_block_txs_by_number};
+use crate::v0_4_0::block::{get_block_header_by_number, get_block_number};
+use crate::v0_4_0::error::JsonRpcError;
+use crate::v0_4_0::transaction::{get_block_tx_hashes_by_number, get_block_txs_by_number};
 use crate::{
-    get_block_number, get_block_status, get_latest_block_number, internal_server_error,
-    ContinuationTokenAsStruct,
+    get_block_status, get_latest_block_number, internal_server_error, ContinuationTokenAsStruct,
 };
 
 /// Rpc server.
